finished api endpoint for updating user info
